refactor(moe): change topk_ids dtype from int32 to int16
Update the data type of topk_ids from int32 to int16 in Kunpeng MoE implementation to reduce memory usage. This change is applied across the Python layer (topk.py and dispatcher) and the C++ kernel (kunpeng_moe.cpp), including the load balancing function which now expects int16 input.

+++ python/sglang/srt/layers/moe/token_dispatcher/kunpeng.py
@@ -525,7 +525,7 @@ class KunpengDispatcher(BaseDispatcher):
         # Build topk_ids_index and topk_weights per TP rank, then allgather across ATTN_TP
         t_topk_allg_start = time.perf_counter()
         if _tp_count > 0:
-            state.topk_ids_index_buf[:batch_size, 0::2] = topk_ids.to(torch.int16)
+            state.topk_ids_index_buf[:batch_size, 0::2] = topk_ids
             state.topk_weights_buf[:batch_size].copy_(topk_weights)
 
         t_topk_allg_end = time.perf_counter()
